Create directory: db_interface.
Replace db_interface.py with db_interface directory.
Divide the content into two .py files.
Break id/name searching for survey into two functions/apis.
